✏️ Fix the duplicated route path warning at the all versions page.
